requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via
    #   -r requirements.in
    #   datasets
    #   fsspec
    #   langchain
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
annotated-types==0.5.0
    # via
    #   -r requirements.in
    #   pydantic
async-timeout==4.0.3
    # via
    #   -r requirements.in
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   jsonlines
certifi==2023.7.22
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.2.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   -r requirements.in
    #   nltk
cmake==3.27.5
    # via
    #   -r requirements.in
    #   triton
dataclasses-json==0.6.0
    # via
    #   -r requirements.in
    #   langchain
datasets==2.14.5
    # via -r requirements.in
dill==0.3.7
    # via
    #   -r requirements.in
    #   datasets
    #   multiprocess
faiss-cpu==1.7.4
    # via -r requirements.in
filelock==3.12.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.4.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.6.0
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
greenlet==2.0.2
    # via
    #   -r requirements.in
    #   sqlalchemy
huggingface-hub==0.17.2
    # via
    #   -r requirements.in
    #   datasets
    #   sentence-transformers
    #   transformers
idna==3.4
    # via
    #   -r requirements.in
    #   requests
    #   yarl
instructorembedding==1.0.1
    # via -r requirements.in
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   torch
joblib==1.3.2
    # via
    #   -r requirements.in
    #   nltk
    #   scikit-learn
jsonlines==4.0.0
    # via -r requirements.in
langchain==0.0.298
    # via -r requirements.in
langsmith==0.0.39
    # via
    #   -r requirements.in
    #   langchain
lit==16.0.6
    # via
    #   -r requirements.in
    #   triton
markdown-it-py==3.0.0
    # via
    #   -r requirements.in
    #   rich
markupsafe==2.1.3
    # via
    #   -r requirements.in
    #   jinja2
marshmallow==3.20.1
    # via
    #   -r requirements.in
    #   dataclasses-json
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
multidict==6.0.4
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via
    #   -r requirements.in
    #   datasets
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   typing-inspect
networkx==3.1
    # via
    #   -r requirements.in
    #   torch
nltk==3.8.1
    # via
    #   -r requirements.in
    #   sentence-transformers
numexpr==2.8.6
    # via
    #   -r requirements.in
    #   langchain
numpy==1.26.0
    # via
    #   -r requirements.in
    #   datasets
    #   langchain
    #   numexpr
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
nvidia-cublas-cu11==11.10.3.66
    # via
    #   -r requirements.in
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via
    #   -r requirements.in
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via
    #   -r requirements.in
    #   torch
nvidia-cuda-runtime-cu11==11.7.99
    # via
    #   -r requirements.in
    #   torch
nvidia-cudnn-cu11==8.5.0.96
    # via
    #   -r requirements.in
    #   torch
nvidia-cufft-cu11==10.9.0.58
    # via
    #   -r requirements.in
    #   torch
nvidia-curand-cu11==10.2.10.91
    # via
    #   -r requirements.in
    #   torch
nvidia-cusolver-cu11==11.4.0.1
    # via
    #   -r requirements.in
    #   torch
nvidia-cusparse-cu11==11.7.4.91
    # via
    #   -r requirements.in
    #   torch
nvidia-nccl-cu11==2.14.3
    # via
    #   -r requirements.in
    #   torch
nvidia-nvtx-cu11==11.7.91
    # via
    #   -r requirements.in
    #   torch
packaging==23.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   marshmallow
    #   transformers
pandas==2.1.1
    # via
    #   -r requirements.in
    #   datasets
pillow==10.0.1
    # via
    #   -r requirements.in
    #   torchvision
pyarrow==8.0.0
    # via
    #   -r requirements.in
    #   datasets
pydantic==2.3.0
    # via
    #   -r requirements.in
    #   langchain
    #   langsmith
pydantic-core==2.6.3
    # via
    #   -r requirements.in
    #   pydantic
pygments==2.16.1
    # via
    #   -r requirements.in
    #   rich
pypdf==3.16.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   langchain
    #   transformers
regex==2023.8.8
    # via
    #   -r requirements.in
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   -r requirements.in
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   torchvision
    #   transformers
rich==13.5.3
    # via -r requirements.in
scikit-learn==1.3.1
    # via
    #   -r requirements.in
    #   sentence-transformers
scipy==1.11.2
    # via
    #   -r requirements.in
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via -r requirements.in
sentencepiece==0.1.99
    # via
    #   -r requirements.in
    #   sentence-transformers
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlalchemy==2.0.21
    # via
    #   -r requirements.in
    #   langchain
sympy==1.12
    # via
    #   -r requirements.in
    #   torch
tenacity==8.2.3
    # via
    #   -r requirements.in
    #   langchain
threadpoolctl==3.2.0
    # via
    #   -r requirements.in
    #   scikit-learn
tokenizers==0.12.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.0.1
    # via
    #   -r requirements.in
    #   sentence-transformers
    #   torchvision
    #   triton
torchvision==0.15.2
    # via
    #   -r requirements.in
    #   sentence-transformers
tqdm==4.66.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   nltk
    #   sentence-transformers
    #   transformers
transformers==4.20.0
    # via
    #   -r requirements.in
    #   sentence-transformers
triton==2.0.0
    # via
    #   -r requirements.in
    #   torch
typing-extensions==4.8.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   -r requirements.in
    #   dataclasses-json
tzdata==2023.3
    # via
    #   -r requirements.in
    #   pandas
urllib3==2.0.5
    # via
    #   -r requirements.in
    #   requests
wheel==0.41.2
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
xxhash==3.3.0
    # via
    #   -r requirements.in
    #   datasets
yarl==1.9.2
    # via
    #   -r requirements.in
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
